+++ src/index.ts
@@ -77,8 +77,8 @@ export {
 export { openUrl, openFile } from './utils/open';
 export { default as systemReport } from './utils/systemReport';
 
-export { default as usageData } from './telemetry/telemetry';
-export { type default as UsageDataMetadata } from './telemetry/TelemetryMetadata';
+export { default as telemetry } from './telemetry/telemetry';
+export { type default as TelemetryMetadata } from './telemetry/TelemetryMetadata';
 export { default as classNames } from './utils/classNames';
 export { truncateMiddle } from './utils/truncateMiddle';